+++ tests/Support/Stub/ColumnSchema.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Tests\Support\Stub;
 
-use Yiisoft\Db\Schema\AbstractColumnSchema;
+use Yiisoft\Db\Schema\Column\AbstractColumnSchema;
 
 final class ColumnSchema extends AbstractColumnSchema
 {